requirements.txt
paho-mqtt>=2.0.0
firebase-admin>=6.0.0
groq>=0.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0